# docs/swagger.yaml
basePath: /api/
definitions:
...
  domain.CreateProductKeluarRequest:
    properties:
      category_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      qty:
        type: string
    required:
    - category_id
    - product_id
    - qty
    type: object
...
  domain.CreateProductRequest:
    properties:
      category_id:
        type: string
      image:
        type: string
      name:
        type: string
      qty:
        type: string
    required:
    - category_id
    - image
    - name
    - qty
    type: object
...
  domain.ErrorMessage:
    properties:
      error:
        type: boolean
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  domain.UpdateProductKeluarRequest:
    properties:
      category_id:
        type: string
      id:
        type: string
      product_id:
        type: string
      qty:
        type: string
    required:
    - category_id
    - product_id
    - qty
    type: object
...
  domain.UpdateProductRequest:
    properties:
      category_id:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      qty:
        type: string
    required:
    - category_id
    - image
    - name
    - qty
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: REST API for FiberInventory
  title: FiberInventory
  version: "1.0"
paths:
...
  /product:
    get:
      description: Retrieve the results for each product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get product results
      tags:
      - Product
  /product/{id}:
    delete:
      description: Delete a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      description: Retrieve the result for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get product result
      tags:
      - Product
    put:
      description: Update a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
  /product/create:
    post:
      consumes:
      - application/json
      description: Create Product
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: createProduct to the application
      tags:
      - product
  /product/hello:
    get:
      description: Return a greeting message to the product
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Greet the Product
      tags:
      - Product
  /productkeluar:
    get:
      description: Retrieve the results for all ProductKeluar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get ProductKeluar results
      tags:
      - ProductKeluar
  /productkeluar/{id}:
    delete:
      description: Delete a specific ProductKeluar
      parameters:
      - description: ProductKeluar ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete a ProductKeluar
      tags:
      - ProductKeluar
    get:
      description: Retrieve the result for a specific ProductKeluar
      parameters:
      - description: ProductKeluar ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get ProductKeluar result
      tags:
      - ProductKeluar
    put:
      description: Update a specific ProductKeluar
      parameters:
      - description: ProductKeluar ID
        in: path
        name: id
        required: true
        type: string
      - description: ProductKeluar Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductKeluarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update a ProductKeluar
      tags:
      - ProductKeluar
  /productkeluar/create:
    post:
      consumes:
      - application/json
      description: Create ProductKeluar
      parameters:
      - description: ProductKeluar information
        in: body
        name: productkeluar
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductKeluarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Create a new ProductKeluar
      tags:
      - ProductKeluar
...
securityDefinitions:
  BearerAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"